desktop/core/requirements.txt
asn1crypto==0.24.0
avro-python3==1.8.2
Babel==2.5.1
BabelDjango==0.2.2
boto==2.46.1
celery[redis]==4.4.5  # For Python 3.8
cffi==1.13.2
channels==2.1.6
channels-redis==2.3.2
configobj==5.0.6
cryptography==3.2
django-auth-ldap==1.3.0
Django==1.11.29 # Django 2 then 3?
django-axes==2.2.0
django-celery-beat==1.4.0
django_celery_results==1.0.4
django-crequest==2018.5.11
django-debug-panel==0.8.3
django-debug-toolbar==1.9.1
django-extensions==1.8.0
django-nose==1.4.5
django_opentracing==1.1.0
django_prometheus==1.0.15
django-webpack-loader==0.5.0
eventlet==0.24.1
future==0.18.2
greenlet==0.4.15
gunicorn==19.9.0
jaeger-client==4.3.0
jdcal==1.0.1
kazoo==2.8.0
kerberos==1.3.0
lockfile==0.12.2
lxml==4.5.0
Mako==1.1.2
Markdown==3.1
mysqlclient==1.4.6
nose==1.3.7
openpyxl==2.6.2
phoenixdb==1.0.0
pyformance==0.3.2
pylint==2.6.0
pylint-django==2.3.0
pysaml2>=4.5.0
pytest==6.0.2
pytest-django==3.10.0
python-dateutil==2.4.2
python-daemon==2.2.4
python-ldap==3.1.0
python-oauth2==1.1.0
pytidylib==0.3.2
pytz==2015.2
pyyaml>=4.2b1
requests>=2.20.0
requests-kerberos==0.12.0
rsa==3.4.2
sasl==0.2.1  # Move to https://pypi.org/project/sasl3/ ?
SQLAlchemy==1.3.8
sqlparse==0.4.1
tablib==0.13.0
thrift==0.13.0
thrift-sasl==0.4.2
pycrypto==2.6.1
jwcrypto==0.6.0
python-jose==3.0.1

